requirements.txt
Django>=4.0.1,<4.1.0
djangorestframework>=3.13.1,<3.14.0
psycopg2>=2.9.2,<2.10.0
Pillow>=9.0.0,<9.1.0
flake8>=4.0.1,<4.1.0
numpy>=1.22.0<1.23.0
